+++ src/claims.rs
@@ -165,7 +165,7 @@ mod serde_timestamp {
     impl<'de> Visitor<'de> for TimestampVisitor {
         type Value = DateTime<Utc>;
 
-        fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+        fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
             formatter.write_str("UTC timestamp")
         }