+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ## [Unreleased]
 
+### Changed
+
+- Update dependencies.
+
 ## 0.2.0 - 2020-05-11
 
 ### Changed