+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ## [Unreleased]
 
+## 0.3.0-beta.1 - 2020-11-08
+
 ### Added
 
 - Add signature getters for untrusted and trusted tokens.